+++ propagator/opentelemetry-propagator-ot-trace/setup.cfg
@@ -47,3 +47,7 @@ where = src
 
 [options.extras_require]
 test =
+
+[options.entry_points]
+opentelemetry_propagator =
+    ottrace = opentelemetry.propagators.OTTracePropagator
